270_nsp/web-search-agent/requirements.txt
python-dotenv>=1.0.0
azure-identity
agent-framework-azure-ai
azure-ai-agentserver-core
azure-ai-agentserver-agentframework

# azure-ai-agentserver-agentframework==1.0.0b7
# azure-ai-agentserver-core==1.0.0b7
# agent-framework==1.0.0b260116
# agent-framework-azure-ai==1.0.0b260116
# python-dotenv>=1.1.1
# azure-monitor-opentelemetry==1.8.1
